requirements.txt
# Cyborg Advisor - Neurosymbolic Rebalancer
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=8.0.0
